Revert "Allow Flutter golden file tests to be flaky (#114450)"
This reverts commit 53e6876226a7729531be7be4fa7321363c5a3b30.
