+++ dev/automated_tests/flutter_test/flutter_gold_test.dart
@@ -8,7 +8,7 @@ import 'dart:typed_data';
 
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
-import 'package:flutter_goldens/src/flutter_goldens_io.dart';
+import 'package:flutter_goldens/flutter_goldens.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:platform/platform.dart';
