--- dev/bots/test/analyze-test-input/root/packages/foo/flaky_golden_ignore.dart
@@ -1,16 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-// This would fail analysis, but it is ignored
-// flutter_ignore_for_file: golden_tag (see analyze.dart)
-
-@Tags(<String>['some-other-tag'])
-
-import 'package:test/test.dart';
-
-import 'golden_class.dart';
-
-void main() {
-  expectFlakyGolden('key', 'String');
-}
