--- dev/bots/test/analyze-test-input/root/packages/foo/flaky_golden_missing_tag.dart
@@ -1,14 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-// The reduced test set tag is missing. This should fail analysis.
-@Tags(<String>['some-other-tag'])
-
-import 'package:test/test.dart';
-
-import 'golden_class.dart';
-
-void main() {
-  expectFlakyGolden('finder', 'missing_tag.png');
-}
