+++ dev/bots/test/analyze-test-input/root/packages/foo/golden_doc.dart
@@ -35,11 +35,8 @@
 /// ```
 /// {@end-tool}
 ///
-/// expectFlakyGolden(a, b)
 
 // Other comments
 // matchesGoldenFile('comment.png');
-// expectFlakyGolden(a, b);
 
 String literal = 'matchesGoldenFile()'; // flutter_ignore: golden_tag (see analyze.dart)
-String flakyLiteral = 'expectFlakyGolden';
