+++ examples/api/test/flutter_test_config.dart
@@ -5,7 +5,7 @@
 import 'dart:async';
 
 
-import 'package:flutter_goldens/flutter_goldens.dart' as flutter_goldens;
+import 'goldens_io.dart' if (dart.library.html) 'goldens_web.dart' as flutter_goldens;
 
 Future<void> testExecutable(FutureOr<void> Function() testMain) {
   // Enable golden file testing using Skia Gold.
