+++ packages/flutter/test/_goldens_io.dart
@@ -2,8 +2,4 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-/// Mack class definition for testing.
-/// Should not be tag-enforced, no analysis failures should be found.
-void matchesGoldenFile(Object key) {
-  return;
-}
+export 'package:flutter_goldens/flutter_goldens.dart' show testExecutable;
