+++ packages/flutter/test/cupertino/date_picker_test.dart
@@ -17,11 +17,10 @@ import 'dart:ui';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:flutter_goldens/flutter_goldens.dart' show expectFlakyGolden;
 import 'package:flutter_test/flutter_test.dart';
 
 // TODO(yjbanov): on the web text rendered with perspective produces flaky goldens: https://github.com/flutter/flutter/issues/110785
-const bool perspectiveTestIsFlaky = isBrowser;
+const bool skipPerspectiveTextGoldens = isBrowser;
 
 // A number of the hit tests below say "warnIfMissed: false". This is because
 // the way the CupertinoPicker works, the hits don't actually reach the labels,
@@ -1198,13 +1197,7 @@ void main() {
       }
 
       await tester.pumpWidget(buildApp(CupertinoDatePickerMode.time));
-
-      if (perspectiveTestIsFlaky) {
-        await expectFlakyGolden(
-          find.byType(CupertinoDatePicker),
-          'date_picker_test.time.initial.png',
-        );
-      } else {
+      if (!skipPerspectiveTextGoldens) {
         await expectLater(
           find.byType(CupertinoDatePicker),
           matchesGoldenFile('date_picker_test.time.initial.png'),
@@ -1212,13 +1205,7 @@ void main() {
       }
 
       await tester.pumpWidget(buildApp(CupertinoDatePickerMode.date));
-
-      if (perspectiveTestIsFlaky) {
-        await expectFlakyGolden(
-          find.byType(CupertinoDatePicker),
-          'date_picker_test.date.initial.png',
-        );
-      } else {
+      if (!skipPerspectiveTextGoldens) {
         await expectLater(
           find.byType(CupertinoDatePicker),
           matchesGoldenFile('date_picker_test.date.initial.png'),
@@ -1226,13 +1213,7 @@ void main() {
       }
 
       await tester.pumpWidget(buildApp(CupertinoDatePickerMode.dateAndTime));
-
-      if (perspectiveTestIsFlaky) {
-        await expectFlakyGolden(
-          find.byType(CupertinoDatePicker),
-          'date_picker_test.datetime.initial.png',
-        );
-      } else {
+      if (!skipPerspectiveTextGoldens) {
         await expectLater(
           find.byType(CupertinoDatePicker),
           matchesGoldenFile('date_picker_test.datetime.initial.png'),
@@ -1243,12 +1224,7 @@ void main() {
       await tester.drag(find.text('4'), Offset(0, _kRowOffset.dy / 2), warnIfMissed: false); // see top of file
       await tester.pump();
 
-      if (perspectiveTestIsFlaky) {
-        await expectFlakyGolden(
-          find.byType(CupertinoDatePicker),
-          'date_picker_test.datetime.drag.png',
-        );
-      } else {
+      if (!skipPerspectiveTextGoldens) {
         await expectLater(
           find.byType(CupertinoDatePicker),
           matchesGoldenFile('date_picker_test.datetime.drag.png'),
@@ -1338,12 +1314,7 @@ void main() {
       ),
     );
 
-    if (perspectiveTestIsFlaky) {
-      await expectFlakyGolden(
-        find.byType(CupertinoTimerPicker),
-        'timer_picker_test.datetime.initial.png',
-      );
-    } else {
+    if (!skipPerspectiveTextGoldens) {
       await expectLater(
         find.byType(CupertinoTimerPicker),
         matchesGoldenFile('timer_picker_test.datetime.initial.png'),
@@ -1354,12 +1325,7 @@ void main() {
     await tester.drag(find.text('59'), Offset(0, _kRowOffset.dy / 2), warnIfMissed: false); // see top of file
     await tester.pump();
 
-    if (perspectiveTestIsFlaky) {
-      await expectFlakyGolden(
-        find.byType(CupertinoTimerPicker),
-        'timer_picker_test.datetime.drag.png',
-      );
-    } else {
+    if (!skipPerspectiveTextGoldens) {
       await expectLater(
         find.byType(CupertinoTimerPicker),
         matchesGoldenFile('timer_picker_test.datetime.drag.png'),
