+++ packages/flutter_goldens_client/lib/skia_client.dart
@@ -8,7 +8,6 @@ import 'dart:io' as io;
 import 'package:crypto/crypto.dart';
 import 'package:file/file.dart';
 import 'package:file/local.dart';
-import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
 import 'package:platform/platform.dart';
 import 'package:process/process.dart';
@@ -139,7 +138,7 @@ class SkiaGoldClient {
   /// The `imgtest` command collects and uploads test results to the Skia Gold
   /// backend, the `init` argument initializes the current test. Used by the
   /// [FlutterPostSubmitFileComparator].
-  Future<void> imgtestInit({ bool isFlaky = false }) async {
+  Future<void> imgtestInit() async {
     // This client has already been initialized
     if (_initialized) {
       return;
@@ -148,7 +147,7 @@ class SkiaGoldClient {
     final File keys = workDirectory.childFile('keys.json');
     final File failures = workDirectory.childFile('failures.json');
 
-    await keys.writeAsString(getKeysJSON(isFlaky: isFlaky));
+    await keys.writeAsString(_getKeysJSON());
     await failures.create();
     final String commitHash = await _getCurrentCommit();
 
@@ -200,7 +199,7 @@ class SkiaGoldClient {
   ///
   /// The [testName] and [goldenFile] parameters reference the current
   /// comparison being evaluated by the [FlutterPostSubmitFileComparator].
-  Future<bool> imgtestAdd(String testName, File goldenFile, { bool isFlaky = false }) async {
+  Future<bool> imgtestAdd(String testName, File goldenFile) async {
     final List<String> imgtestCommand = <String>[
       _goldctl,
       'imgtest', 'add',
@@ -210,7 +209,7 @@ class SkiaGoldClient {
       '--test-name', cleanTestName(testName),
       '--png-file', goldenFile.path,
       '--passfail',
-      ..._getPixelMatchingArguments(isFlaky: isFlaky),
+      ..._getPixelMatchingArguments(),
     ];
 
     final io.ProcessResult result = await process.run(imgtestCommand);
@@ -260,7 +259,7 @@ class SkiaGoldClient {
   /// The `imgtest` command collects and uploads test results to the Skia Gold
   /// backend, the `init` argument initializes the current tryjob. Used by the
   /// [FlutterPreSubmitFileComparator].
-  Future<void> tryjobInit({ bool isFlaky = false }) async {
+  Future<void> tryjobInit() async {
     // This client has already been initialized
     if (_tryjobInitialized) {
       return;
@@ -269,7 +268,7 @@ class SkiaGoldClient {
     final File keys = workDirectory.childFile('keys.json');
     final File failures = workDirectory.childFile('failures.json');
 
-    await keys.writeAsString(getKeysJSON(isFlaky: isFlaky));
+    await keys.writeAsString(_getKeysJSON());
     await failures.create();
     final String commitHash = await _getCurrentCommit();
 
@@ -324,7 +323,7 @@ class SkiaGoldClient {
   ///
   /// The [testName] and [goldenFile] parameters reference the current
   /// comparison being evaluated by the [FlutterPreSubmitFileComparator].
-  Future<void> tryjobAdd(String testName, File goldenFile, { bool isFlaky = false}) async {
+  Future<void> tryjobAdd(String testName, File goldenFile) async {
     final List<String> imgtestCommand = <String>[
       _goldctl,
       'imgtest', 'add',
@@ -333,7 +332,7 @@ class SkiaGoldClient {
         .path,
       '--test-name', cleanTestName(testName),
       '--png-file', goldenFile.path,
-      ..._getPixelMatchingArguments(isFlaky: isFlaky),
+      ..._getPixelMatchingArguments(),
     ];
 
     final io.ProcessResult result = await process.run(imgtestCommand);
@@ -363,42 +362,6 @@ class SkiaGoldClient {
     }
   }
 
-  List<String> _getPixelMatchingArguments({ required bool isFlaky }) {
-    if (isFlaky) {
-      return _getFlakyPixelMatchingArguments();
-    } else {
-      return _getNormalPixelMatchingArguments();
-    }
-  }
-
-  List<String> _getFlakyPixelMatchingArguments() {
-    // The algorithm to be used when matching images. The available options are:
-    // - "fuzzy": Allows for customizing the thresholds of pixel differences.
-    // - "sobel": Same as "fuzzy" but performs edge detection before performing
-    //            a fuzzy match.
-    const String algorithm = 'fuzzy';
-
-    // The number of pixels in this image that are allowed to differ from the
-    // baseline.
-    //
-    // The chosen number - 1 billion - indicates that a flaky test should pass
-    // no matter how many pixels are different from the master golden.
-    const int maxDifferentPixels = 1000 * 1000 * 1000;
-
-    // The maximum acceptable difference per pixel.
-    //
-    // The chosen number - 1020 - is the maximum supported pixel delta and
-    // indicates that a flaky test should pass no matter how far the new pixels
-    // deviate from the master golden.
-    const int pixelDeltaThreshold = 1020;
-
-    return <String>[
-      '--add-test-optional-key', 'image_matching_algorithm:$algorithm',
-      '--add-test-optional-key', 'fuzzy_max_different_pixels:$maxDifferentPixels',
-      '--add-test-optional-key', 'fuzzy_pixel_delta_threshold:$pixelDeltaThreshold',
-    ];
-  }
-
   // Constructs arguments for `goldctl` for controlling how pixels are compared.
   //
   // For AOT and CanvasKit exact pixel matching is used. For the HTML renderer
@@ -406,7 +369,7 @@ class SkiaGoldClient {
   // because Chromium cannot exactly reproduce the same golden on all computers.
   // It seems to depend on the hardware/OS/driver combination. However, those
   // differences are very small (typically not noticeable to human eye).
-  List<String> _getNormalPixelMatchingArguments() {
+  List<String> _getPixelMatchingArguments() {
     // Only use fuzzy pixel matching in the HTML renderer.
     if (!_isBrowserTest || _isBrowserCanvasKitTest) {
       return const <String>[];
@@ -522,17 +485,17 @@ class SkiaGoldClient {
   /// Currently, the only key value pairs being tracked is the platform the
   /// image was rendered on, and for web tests, the browser the image was
   /// rendered on.
-  @visibleForTesting
-  String getKeysJSON({ bool isFlaky = false}) {
+  String _getKeysJSON() {
     final Map<String, dynamic> keys = <String, dynamic>{
       'Platform' : platform.operatingSystem,
       'CI' : 'luci',
-      'markedFlaky' : isFlaky.toString(),
     };
     if (_isBrowserTest) {
       keys['Browser'] = _browserKey;
       keys['Platform'] = '${keys['Platform']}-browser';
-      keys['WebRenderer'] = _isBrowserCanvasKitTest ? 'canvaskit' : 'html';
+      if (_isBrowserCanvasKitTest) {
+        keys['WebRenderer'] = 'canvaskit';
+      }
     }
     return json.encode(keys);
   }
