+++ packages/flutter_tools/lib/src/web/bootstrap.dart
@@ -224,8 +224,7 @@ String generateTestEntrypoint({
   Future<void> main() async {
     ui.debugEmulateFlutterTesterEnvironment = true;
     await ui.webOnlyInitializePlatform();
-    webTestUri = Uri.parse('${Uri.file(absolutePath)}');
-    webGoldenComparator = DefaultWebGoldenComparator(webTestUri);
+    webGoldenComparator = DefaultWebGoldenComparator(Uri.parse('${Uri.file(absolutePath)}'));
     (ui.window as dynamic).debugOverrideDevicePixelRatio(3.0);
     (ui.window as dynamic).webOnlyDebugPhysicalSizeOverride = const ui.Size(2400, 1800);
